rename right side bar
