+++ src/layouts/CourseListSideBar/CourseDetails.tsx
