+++ src/layouts/DarsPanel/DarsPanel.tsx
@@ -1,0 +1,5 @@
+export const DarsPanel: React.FC = () => {
+  return <div>Left Panel</div>;
+};
+
+export default DarsPanel;
