--- src/layouts/DarsSideBar/index.ts
@@ -1,1 +1,0 @@
-export { default } from './DarsSideBar';
