--- src/layouts/LeftPanel/LeftPanel.tsx
@@ -1,5 +1,0 @@
-export const LeftPanel: React.FC = () => {
-  return <div>Left Panel</div>;
-};
-
-export default LeftPanel;
