--- src/layouts/LeftPanel/index.ts
@@ -1,1 +1,0 @@
-export { default } from './LeftPanel';
